Add files via upload
This script uses perl to perform a point_energy_calculation_on_multiple_proteins in a foreach loop with system calls. 
The .mdp file is a Gromacs type molecular dynamics parameter file which includes the group scheme performing a zero time, zero kinetic energy calculation of unsolvated proteins in a vacuum, that is the initial conditions.